Debug code added
Added print lines to discover the cause of the slide not working. I seemed to figure it out: The slide in question that doesn't work fails because the previous slide was canceled prematurely (by jumping, falling, ascending, or grapple) however, the delay that allows the animation to complete still gets called and has the player uncrouch almost the instant they crouch when starting the second slide. However I'm not sure how to fix it.